refactor: エントリースクリプトを src/entry に整理
