+++ index.html
@@ -88,9 +88,9 @@
     }
   }
   </script>
-  <script type="module" src="./main.js"></script>
-  <script type="module" src="./bg.js"></script>
-  <script type="module" src="./works.js"></script>
+  <script type="module" src="./src/entry/avatar.js"></script>
+  <script type="module" src="./src/entry/background.js"></script>
+  <script type="module" src="./src/entry/works-gallery.js"></script>
 </body>
 
 </html>
